Added file transfer files
You can use these to transfer files from one computer or pc to a computer or pc. 
Note: Should be on the same network. The phone to pc was made precisely for transferring photos from phone to pc, and the FileTransfer was made precisely for transfering files from one pc to another, but even this can be used to trasnfer from phone to pc